requirements.txt
confluent-kafka==2.13.0
requests==2.31.0
snowflake-connector-python==3.12.0
python-dotenv==1.0.0
